refactored movementLabel into labels.tsx
